% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assigntoclusters.R
\name{assignToClusters_interface}
\alias{assignToClusters_interface}
\alias{assign_to_clusters}
\title{Assign data points to clusters based on centroids}
\usage{
assign_to_clusters(pointX, pointY, centroidsX, centroidsY, clusterAssignments)
}
\arguments{
\item{pointX}{A numeric vector representing the x-coordinates of data points.}

\item{pointY}{A numeric vector representing the y-coordinates of data points.}

\item{centroidsX}{A numeric vector representing the x-coordinates of centroids.}

\item{centroidsY}{A numeric vector representing the y-coordinates of centroids.}

\item{clusterAssignments}{A vector of integers indicating the cluster assignments for each data point.}
}
\value{
A vector of integers representing the cluster assignments for each data point.
}
\description{
This function assigns data points to clusters based on the given centroids.
}
\examples{
pointX <- c(1.0, 2.0, 3.0, 4.0, 5.0)
pointY <- c(2.0, 3.0, 4.0, 5.0, 6.0)
centroidsX <- c(1.5, 4.0)
centroidsY <- c(2.5, 5.0)
clusterAssignments <- integer(length(pointX))
assign_to_clusters(pointX, pointY, centroidsX, centroidsY, clusterAssignments)
}
