% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculatedistance.R
\name{calculateDistance_interface}
\alias{calculateDistance_interface}
\alias{calculate_distance}
\title{Calculate the Euclidean distance between two points}
\usage{
calculate_distance(x1, y1, x2, y2)
}
\arguments{
\item{x1}{A numeric value representing the x-coordinate of the first point.}

\item{y1}{A numeric value representing the y-coordinate of the first point.}

\item{x2}{A numeric value representing the x-coordinate of the second point.}

\item{y2}{A numeric value representing the y-coordinate of the second point.}
}
\value{
The Euclidean distance between the two points as a numeric value.
}
\description{
This function calculates the Euclidean distance between two points represented by their coordinates (x1, y1) and (x2, y2).
}
\examples{
x1 <- 1.0
y1 <- 2.0
x2 <- 3.0
y2 <- 4.0
distance <- calculate_distance(x1, y1, x2, y2)
}
