% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmm.R
\name{fitGMM_interface}
\alias{fitGMM_interface}
\alias{gmm_algo}
\title{Gaussian Mixture Model Algorithm}
\usage{
gmm_algo(data, numComponents, max_iterations)
}
\arguments{
\item{data}{The input data matrix or data frame.}

\item{numComponents}{The number of Gaussian components to fit in the GMM.}

\item{max_iterations}{The maximum number of iterations for the EM algorithm.}
}
\value{
A list containing the following components:
  - 'mu' - A matrix of estimated means for each Gaussian component.
  - 'sigma' - A list of estimated covariance matrices for each component.
  - 'pi' - A vector of estimated mixing proportions for each component.
  - 'log_likelihood' - The log-likelihood of the final GMM model.
}
\description{
This function fits a Gaussian Mixture Model (GMM) to the input data using the Expectation-Maximization (EM) algorithm.
}
\examples{
data <- c(1.0, 2.0, 3.0, 4.0, 5.0)
numComponents <- 2
maxIters <- 100
gmm_algo(data, numComponents, maxIters)
}
