% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans.R
\name{kMeans_interface}
\alias{kMeans_interface}
\alias{kmeans_algo}
\title{Perform K-means Clustering}
\usage{
kmeans_algo(pointX, pointY, k, maxIterations, centroidsX, centroidsY)
}
\arguments{
\item{pointX}{A numeric vector representing the x-coordinates of data points.}

\item{pointY}{A numeric vector representing the y-coordinates of data points.}

\item{k}{An integer specifying the number of clusters to create.}

\item{maxIterations}{An integer specifying the maximum number of iterations for the K-means algorithm.}

\item{centroidsX}{A numeric vector representing the initial x-coordinates of cluster centroids.}

\item{centroidsY}{A numeric vector representing the initial y-coordinates of cluster centroids.}
}
\value{
A numeric vector containing cluster assignments for each data point. Each element of the vector corresponds to the cluster ID to which the corresponding data point belongs.
}
\description{
This function performs K-means clustering on a set of data points in two-dimensional space.
}
\examples{
pointX <- c(1.0, 2.0, 3.0)
pointY <- c(4.0, 5.0, 6.0)
k <- 3
maxIterations <- 100
centroidsX <- numeric(k)
centroidsY <- numeric(k)
kmeans_algo(pointX, pointY, k, maxIterations, centroidsX, centroidsY)

}
